refactor(billing): extract subscription panels
